api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==11.0.0
python-multipart==0.0.9
torch==2.8.0
torchvision==0.23.0

